requirements.txt
numpy==1.26.1
pandas==2.0.3
matplotlib==3.8.1
pvlib==0.9.5 #aangepast
paho-mqtt==1.6.1
pypsa==0.25.2
pvfactors==1.5.2
